# app/utils/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Estes são os serviços disponíveis da nossa API de e-commerce. <br>Mais informações podem ser encontradas em [http://upf.br/](http://upf.br).",
    "version" : "1.0.0",
    "title" : "eCommerce API",
    "termsOfService" : "http://localhost:3000/ecommerce/termos.html",
    "contact" : {
      "email" : "apiecommerce@upf.br"
    }
  },
  "host" : "localhost:3000",
  "basePath" : "/ecommerce",
...
  "schemes" : [ "http" ],
  "paths" : {
    "/cupons" : {
      "get" : {
        "tags" : [ "cupomdesconto" ],
        "summary" : "Retorna todos os cupons de desconto.",
        "operationId" : "listarTudo",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Sucesso"
          },
          "502" : {
            "description" : "Erro ao recuperar dados"
          }
        },
        "security" : [ {
          "ecommerce_auth" : [ "read:cupons" ]
        } ]
      },
      "post" : {
        "tags" : [ "cupomdesconto" ],
        "summary" : "Adicione um novo cupom de desconto",
        "description" : "",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "cupomDesconto",
          "description" : "Objeto Cupom de Desconto que será adicionado.",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/CupomDesconto"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Cupom de desconto criado"
          },
          "400" : {
            "description" : "Valores inválidos de entrada"
          },
          "500" : {
            "description" : "Não foi possível executar a requisição"
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "ecommerce_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "http://localhost:3000/ecommerce/auth/google",
      "flow" : "implicit",
      "scopes" : {
        "write:cupons" : "modificar cupons",
        "read:cupons" : "visualizar cupons"
      }
    },
    "api_key" : {
      "type" : "apiKey",
      "name" : "authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "CupomDesconto" : {
      "type" : "object",
      "required" : [ "percentualDesconto" ],
      "properties" : {
        "dataInicial" : {
          "type" : "string",
          "format" : "date"
        },
        "dataFinal" : {
          "type" : "string",
          "format" : "date"
        },
        "valorInicial" : {
          "type" : "number",
          "example" : 99.9
        },
        "valorFinal" : {
          "type" : "number"
        },
        "tipo" : {
          "type" : "string",
          "description" : "Tipo do Cupom de Desconto",
          "enum" : [ "DESCONTO", "FRETEGRATIS", "PROMOCIONAL" ]
        },
        "quantidadeCupons" : {
          "type" : "integer",
          "example" : 500
        },
        "quantidadeUsada" : {
          "type" : "integer"
        },
        "percentualDesconto" : {
          "type" : "number",
          "example" : 10.0
        }
      }
    },
...
  }
}